Refactor Code

+++ app/Http/Controllers/V1/RestaurantController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers\V1;
 
+use App\Models\User;
+use App\Models\Restaurant;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Models\Restaurant;
+use App\Models\RestaurantPicture;
+use Illuminate\Support\Facades\Storage;
 
 class RestaurantController extends Controller
 {
@@ -62,22 +65,42 @@ class RestaurantController extends Controller
     {
         $this->validate($request, [
             'user_id'             => 'required|integer|exists:users,id',
-            'cousine_type_id'     => 'required|integer|exists:cousine_types,id',
+            'cousine_type_id.*'   => 'required|integer|exists:cousine_types,id',
             'name'                => 'required|alpha|max:20',
             'address1'            => 'required|string|max:50',
             'address2'            => 'nullable|string|max:50',
             'zip_code'            => 'required|integer|min:6',
             'phone'               => 'required|integer|min:10',
-            'profile_picture'     => 'required|mimes:jpg,jpeg,png,bmp,tiff'
+            'logo'                => 'required|mimes:jpg,jpeg,png,bmp,tiff',
+            'type'                => 'required|in:M,O',
+            'picture.*'           => 'required|mimes:jpg,jpeg,png,bmp,tiff'
+
         ]);
+        $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->logo->getClientOriginalExtension();
+        $request->logo->storeAs("public/pictures", $imageName);
 
-        $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->profile_picture->getClientOriginalExtension();
-        $request->profile_picture->storeAs("public/pictures", $imageName);
+        $user = User::findOrFail($request->user_id);
 
-        $restaurant = Restaurant::create($request->only('cousine_type_id', 'name', 'address1', 'address2', 'phone', 'zip_code') + ['profile_picture' => $imageName]);
-        $restaurant->users()->syncWithoutDetaching($request->user_id);
-
-        return ok('Restaurant created successfully!',  $restaurant->load('users'));
+        $image = array();
+        if ($request->hasFile('picture')) {
+            foreach ($request->picture as $file) {
+                $image_name =  str_replace(".", "", (string)microtime(true)) . '.' . $file->getClientOriginalExtension();
+                $upload_path =  'public/pictures';
+                $file->storeAs($upload_path, $image_name);
+                $image[] = [
+                    'restaurant_id' => $request->id,
+                    'picture' => $image_name,
+                    'type' => $request->type
+                ];
+            }
+        }
+        if ($user->role_id == 2) {
+            $restaurant = Restaurant::create($request->only('name', 'address1', 'address2', 'phone', 'zip_code') + ['logo' => $imageName]);
+            $restaurant->users()->syncWithoutDetaching([$request->user_id => ['is_owner' => true]]);
+            $restaurant->cousines()->syncWithoutDetaching($request->cousine_type_id);
+            $restaurant->pictures()->createMany($image);
+            return ok('Restaurant created successfully!',  $restaurant->load('users', 'pictures', 'cousines'));
+        }
     }
 
     /**
@@ -90,25 +113,48 @@ class RestaurantController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
-            'user_id'             => 'nullable|integer|exists:users,id',
-            'cousine_type_id'     => 'nullable|integer|exists:cousine_types,id',
+            'user_id.*'           => 'nullable|integer|exists:users,id',
+            'cousine_type_id.*'   => 'nullable|integer|exists:cousine_types,id',
             'name'                => 'required|alpha|max:20',
             'address1'            => 'nullable|string|max:50',
             'address2'            => 'nullable|string|max:50',
             'zip_code'            => 'nullable|integer|min:6',
             'phone'               => 'nullable|integer|min:10',
-            'profile_picture'     => 'required|mimes:jpg,jpeg,png,bmp,tiff'
+            'logo'                => 'required|mimes:jpg,jpeg,png,bmp,tiff',
+            'type'                => 'nullable|in:M,O',
+            'picture.*'           => 'nullable|mimes:jpg,jpeg,png,bmp,tiff'
         ]);
 
         $restaurant = Restaurant::findOrFail($id);
+        if ($restaurant->logo) {
+            Storage::delete("public/pictures/" . $restaurant->logo);
+            $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->logo->getClientOriginalExtension();
+            $request->logo->storeAs("public/pictures", $imageName);
+        }
+        $restaurant->update($request->only('name', 'address1', 'address2', 'phone', 'zip_code') + ['logo' => $imageName]);
+        $restaurant->cousines()->sync($request->cousine_type_id);
 
-        $imageName = str_replace(".", " ", (string)microtime(true)) . '.' . $request->profile_picture->getClientOriginalExtension();
-        $request->profile_picture->storeAs("public/pictures", $imageName);
+        $picture = RestaurantPicture::findOrFail($id);
+        if ($picture->picture) {
+            $image = array();
+            if ($request->hasFile('picture')) {
+                Storage::delete("public/pictures/" . $picture->picture);
+                $restaurant->pictures()->delete();
+                foreach ($request->picture as $file) {
+                    $image_name =  str_replace(".", "", (string)microtime(true)) . '.' . $file->getClientOriginalExtension();
+                    $upload_path =  'public/pictures';
+                    $file->storeAs($upload_path, $image_name);
+                    $image[] = [
+                        'restaurant_id' => $request->id,
+                        'picture' => $image_name,
+                        'type' => $request->type
+                    ];
+                }
+            }
+        }
 
-        $restaurant->update($request->only('cousine_type_id', 'name', 'address1', 'address2', 'phone', 'zip_code') + ['profile_picture' => $imageName]);
-        $restaurant->users()->sync($request->user_id);
-
-        return ok('Restaurant updated successfully!',  $restaurant->load('users'));
+        $restaurant->pictures()->createMany($image);
+        return ok('Restaurant updated successfully!', $restaurant->load('cousines', 'pictures'));
     }
 
     /**
@@ -119,7 +165,7 @@ class RestaurantController extends Controller
      */
     public function get($id)
     {
-        $restaurant = Restaurant::with(['users', 'cousines', 'orders'])->findOrFail($id);
+        $restaurant = Restaurant::with(['users', 'cousines', 'pictures', 'orders'])->findOrFail($id);
 
         return ok('Restaurant retrieved successfully',  $restaurant);
     }
@@ -133,8 +179,10 @@ class RestaurantController extends Controller
     public function delete($id)
     {
         $restaurant = Restaurant::findOrFail($id);
-        if ($restaurant->users()->count() > 0) {
+        if ($restaurant->users()->count() > 0 && $restaurant->pictures()->count() > 0 && $restaurant->cousines()->count() > 0) {
             $restaurant->users()->detach();
+            $restaurant->cousines()->detach();
+            $restaurant->pictures()->delete();
         }
         $restaurant->delete();
         return ok('Restaurant deleted successfully');
